requirements.txt
PyQt5
PyQtWebEngine
pandas
webdriver_manager
flask==2.3.2
flask-socketio==5.3.2
asyncio 
uuid 
pypdfium2
selenium==4.21.0
waitress==2.1.2
PyMuPDF
pdfplumber
openpyxl
flask_cors
xlsxwriter
eventlet
python-dotenv
pytest
pytest-flask
flask_sqlalchemy
flask_migrate
coverage
pylint
black